feat: update referenceCalls return structure
