+++ lib/collector.js
@@ -21,6 +21,8 @@ export class Collector {
   }
 
   exportResolversCalls() {
-    return Object.fromEntries(this.resolverCalls)
+    return Array.from(this.resolverCalls).map(([key, entry]) => {
+      return { key, ...entry }
+    })
   }
 }
